moves some internal dirs
