+++ src/details/r3d_half.h
